# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is list API's to be used in Quiz App.",
        "title": "Quiz App API's",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/quizes/get-categories/{user_id}": {
            "get": {
                "description": "This API will provide list of quiz categories",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Quizes"
                ],
                "summary": "This API will provide list of quiz categories",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.categoryInfo"
                        }
                    }
                }
            }
        },
        "/quizes/get-contest-joined-by-user/{user_id}": {
            "get": {
                "description": "This API will list contest joined by user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Quizes"
                ],
                "summary": "This API will list contest joined by user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.JoinedContestResponse"
                        }
                    }
                }
            }
        },
        "/quizes/get-question-by-category/{category_id}": {
            "get": {
                "description": "This API will provide list of quiz questions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Quizes"
                ],
                "summary": "This API will provide list of quiz questions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "quiz category id",
                        "name": "category_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.quizInfo"
                        }
                    }
                }
            }
        },
        "/quizes/user-join-contest": {
            "post": {
                "description": "This API will make user to join a contest",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Quizes"
                ],
                "summary": "This API will make user to join a contest",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id",
                        "name": "user_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "quiz category id",
                        "name": "category_id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.JoinContestResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.ContestInfo": {
            "type": "object",
            "properties": {
                "contest_amount": {
                    "type": "integer",
                    "example": 10000
                },
                "contest_date": {
                    "type": "string",
                    "example": "2024-12-21T18:00:00+05:30"
                },
                "contest_id": {
                    "type": "integer",
                    "example": 1
                },
                "contest_name": {
                    "type": "string",
                    "example": "GK"
                }
            }
        },
        "controllers.JoinContestResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Contest joined successfully."
                },
                "status": {
                    "type": "string",
                    "example": "1"
                }
            }
        },
        "controllers.JoinedContestResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controllers.ContestInfo"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Contest list found."
                },
                "status": {
                    "type": "string",
                    "example": "1"
                }
            }
        },
...
        "controllers.categoryInfo": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "active": {
                                "type": "string",
                                "example": "1"
                            },
                            "created": {
                                "type": "string",
                                "example": "2024-12-17T18:07:19+05:30"
                            },
                            "icon": {
                                "type": "string",
                                "example": "https://quizbuck.s3.ap-south-1.amazonaws.com/uploads/1734090491_new.jpg"
                            },
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "join_amount": {
                                "type": "integer",
                                "example": 100
                            },
                            "num_of_users_can_join": {
                                "type": "integer",
                                "example": 20
                            },
                            "num_of_users_have_joined": {
                                "type": "integer",
                                "example": 0
                            },
                            "quiz_time": {
                                "type": "string",
                                "example": "2024-12-17T18:00:00+05:30"
                            },
                            "title": {
                                "type": "string",
                                "example": "GK"
                            },
                            "total_price": {
                                "type": "integer",
                                "example": 100000
                            },
                            "user_has_joined": {
                                "type": "boolean",
                                "example": true
                            }
                        }
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Category list data found."
                },
                "status": {
                    "type": "string",
                    "example": "1"
                }
            }
        },
...
        "controllers.quizInfo": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "object",
                    "properties": {
                        "category_id": {
                            "type": "integer",
                            "example": 1
                        },
                        "correct_answer": {
                            "type": "string",
                            "example": "c"
                        },
                        "createdAt": {
                            "type": "string"
                        },
                        "id": {
                            "type": "integer",
                            "example": 1
                        },
                        "level": {
                            "type": "string",
                            "example": "easy"
                        },
                        "option_a": {
                            "type": "string",
                            "example": "Unites States of America"
                        },
                        "option_b": {
                            "type": "string",
                            "example": "England"
                        },
                        "option_c": {
                            "type": "string",
                            "example": "India"
                        },
                        "option_d": {
                            "type": "string",
                            "example": "Sri Lanka"
                        },
                        "question": {
                            "type": "string",
                            "example": "Where is Delhi?"
                        }
                    }
                },
                "message": {
                    "type": "string",
                    "example": "New user created"
                },
                "status": {
                    "type": "string",
                    "example": "1"
                }
            }
        },
...
    }
}